Decompile func0f00b510

+++ src/include/constants.h
@@ -14,7 +14,7 @@
 #define ABS(val)              ((val) > 0 ? (val) : -(val))
 #define ALIGN8(val)           (((val) + 0x7 | 0x7) ^ 0x7)
 #define ALIGN16(val)          (((val) + 0xf | 0xf) ^ 0xf)
-#define ALIGN64(val)          (((val) + 0x3f | 0x3f) ^ 0x3f)
+#define ALIGN64(val)          ((((u32)val) + 0x3f | 0x3f) ^ 0x3f)
 #define ARRAYCOUNT(a)         (s32)(sizeof(a) / sizeof(a[0]))
 #define CHRRACE(chr)          (chr ? chr->race : RACE_HUMAN)
 #define FRAMEDURATION         (1 / 60.0f)
